requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
motor==3.5.1
pymongo==4.8.0
python-jose==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
authlib==1.3.1
pydantic-settings==2.4.0
itsdangerous==2.2.0
email-validator==2.2.0
stripe==9.7.0
pytest==8.3.2
pytest-asyncio==0.23.8
slack-sdk==3.33.1
cryptography==42.0.8
msal==1.30.0
google-auth==2.33.0
google-auth-httplib2==0.2.0
fastapi-limiter==0.1.6
redis==5.0.8
httpx==0.27.2
python-multipart==0.0.9
langchain==0.2.16
langgraph==0.2.32
langchain-openai==0.1.23
langchain-google-genai==1.0.10
google-generativeai==0.7.2
asyncpg==0.29.0
pgvector==0.2.5
transformers==4.44.2
huggingface_hub==0.26.2
openai==1.40.6
torch==2.10.0
